wayland: move pointer to its own file
